add: medusa price input
